pyproject.toml
[tool.poetry]
name = "pizzashop-api-python"
version = "0.1.0"
description = "Adaptação da API feita em Bun para rodar em ambientes não UNIX"
authors = ["tuliomitico <tulioneedforspeed@gmail.com>"]
readme = "README.md"
packages = [{include = "pizzashop_api_python"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "~2.2.0"
Werkzeug = "2.3.7"
flask-smorest = "~0.42"
sqlalchemy = "^2.0.27"
psycopg2 = "^2.9.9"
cuid2 = "^2.0.0"
flask-jwt-extended = "^4.6.0"
alembic = "^1.13.1"
yachalk = "^0.1.5"
pytz = "^2024.1"
flask-marshmallow = "^1.2.0"
marshmallow-sqlalchemy = "^1.0.0"
flask-sqlalchemy = "^3.1.1"

[tool.poetry.dev-dependencies]
factory-boy = "^3.3.0"


[tool.poetry.group.dev.dependencies]
faker-commerce = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
